fix: 카테고리 기본키 타입 Byte로 수정

+++ backend/clothes/src/main/java/com/ssafy/ozz/clothes/clothes/dto/request/ClothesCreateRequest.java
@@ -24,7 +24,7 @@ public record ClothesCreateRequest(
         List<Season> seasonList,
         List<Style> styleList,
         List<Pattern> patternList,
-        Long categoryLowId
+        Byte categoryLowId
 ) {
     public Clothes toEntity(CategoryLow categoryLow, Long imageFileId, Long userId) {
         return Clothes.builder()
